plankton: Support drivers for Plankton backend
This commit decouples the Plankton backend from the Pithos backend, by
making the Plankton backend able to use external drivers to handle
images and snapshots.

The Plankton backend is loading and initiating the driver class that is
defined in the new 'PLANKTON_BACKEND_DRIVER' setting and uses this driver
to handle images and snapshots.

The previous implementation of Plankton that was using the Pithos
backend has been ported to the 'PithosPlantonDriver', without any
modification on how this driver works.

The JSONFileBackend has totally been removed, since it was not used and
was not a functional backend implementation. However, this commit allows
to easily implement such a JSON file driver, if needed.
